requirements.txt
Python 3.10
deepface==0.0.78
tensorflow==2.12.0
keras==2.12.0
retina-face==0.0.17
opencv-python
matplotlib
pillow
